client/requirements.txt
PySide6==6.4.2
pycryptodome==3.15.0